requirements.txt
apiclient==1.0.3
click==6.7
Flask==0.12
Flask-WTF==0.14.2
google-api-python-client==1.6.1
gunicorn==19.6.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
oauth2client==4.0.0
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
-e git://github.com/flockchat/pyflock@7983837f23a2cbce10fb61b3546411c9d6d16f32#egg=pyflock
PyJWT==1.4.2
requests==2.12.5
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.20
Werkzeug==0.11.15
WTForms==2.1
